Introduction

YaoLang is a domain specific language (DSL) built based on Julia builtin expression with extended semantic on quantum control, measure and position. Its (extended) syntax is very simple:

Semantics

The semantic of YaoLang tries to make use of Julia semantic as much as possible so you don't feel this is not Julian. But since the quantum circuit has some special semantic that Julia expression cannot express directly, the semantic of Julia expression is extended in YaoLang.

The point of this new IR is it make use of Julia native control flow directly instead of unroll the loop and conditions into a Julia type, such as Chain, Kron, ConditionBlock in QBIR, which improves the performance and provide possibility of further compiler optimization by analysis done on quantum circuit and classical control flows.

Gate Position

gate positions are specific with => at each line, the => operator inside function calls will not be parsed, e.g

1 => H # apply Hadamard gate on the 1st qubit
foo(1=>H) # it means normal Julia pair
1=>foo(x, y, z) # it will parse foo(x, y, z) as a quantum gate/circuit, but will error later if type inference finds they are not.

all the gate or circuit's position should be specified by its complete locations, e.g

1:n => qft(n) # right
1 => qft(n) # wrong

but single qubit gates can use multi-location argument to represent repeated locations, e.g

1:n => H # apply H on 1:n locations

Control

@ctrl is parsed as a keyword (means you cannot overload it) in each program, like QBIR, its first argument is the control location with signs as control configurations and the second argument is a normal gate position argument introduce above.

Measure

@measure is another reserved special function parsed that has specific semantic in the IR (measure the locations passed to it).

Usage

using it is pretty simple, just use @device macro to annotate a "device" function, like CUDA programming, this device function should not return anything but nothing.

The compiler will compile this function definition to a generic circuit Circuit with the same name. A generic circuit is a generic quantum program that can be overload with different Julia types, e.g

@device function qft(n::Int)
    1 => H
    for k in 2:n
        @ctrl k 1=>shift(2π/2^k)
    end

    if n > 1
        2:n => qft(n-1)
    end
end

There is no need to worry about global position: everything can be defined locally and we will infer the correct global location later either in compile time or runtime.

note: all the quantum gates should be annotate with its corresponding locations, or the compiler will not treat it as a quantum gate but instead of the original Julia expression.

Why?

There are a few reasons that we need a fully compiled DSL now.

1. Extensibility

Things in YaoBlocks like

function apply!(r::AbstractRegister, pb::PutBlock{N}) where {N}
    _check_size(r, pb)
    instruct!(r, mat_matchreg(r, pb.content), pb.locs)
    return r
end

# specialization
for G in [:X, :Y, :Z, :T, :S, :Sdag, :Tdag]
    GT = Expr(:(.), :ConstGate, QuoteNode(Symbol(G, :Gate)))
    @eval function apply!(r::AbstractRegister, pb::PutBlock{N,C,<:$GT}) where {N,C}
        _check_size(r, pb)
        instruct!(r, Val($(QuoteNode(G))), pb.locs)
        return r
    end
end

cannot be easily extended without define new dispatch on specialized instruction. Similarly, as long as there is a new instruction in low level, one need to redefine the dispatch in YaoBlocks however this is not necessary!

2. Work with classical computers

Programs defined in such way are just "normal" Julia programs, but quantum devices can be used as accelerator in a similar way comparing to GPU as an optimization.

3. More elegant and better performance

In YaoBlocks, a large quantum circuit can easily lost its structure if it is controlled, unless the programmer specialize the control block manually. Now we can map local locations into its callee location using the brand new API, thus anything in theory is composable can be executed in such way.

API References

YaoLang.H — Constant

H

The Hadamard gate.

Definition

\[\frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1\ 1 & -1 \end{pmatrix}\]

YaoLang.Rx — Constant

Rx(theta::Real)

Return a rotation gate on X axis.

YaoLang.Ry — Constant

Ry(theta::Real)

Return a rotation gate on Y axis.

YaoLang.Rz — Constant

Rz(theta::Real)

Return a rotation gate on Z axis.

YaoLang.T — Constant

T

The T gate.

YaoLang.X — Constant

X

The Pauli X gate.

YaoLang.Y — Constant

Y

The Pauli Y gate.

YaoLang.Z — Constant

Z

The Pauli Z gate.

YaoLang.phase — Constant

phase(theta)

Global phase gate.

Definition

\[exp(iθ) \mathbf{I}\]

YaoLang.rot — Constant

rot(axis, θ::T, m::Int=size(axis, 1)) where {T <: Real}

General rotation gate, axis is the rotation axis, θ is the rotation angle. m is the size of rotation space, default is the size of rotation axis.

YaoLang.shift — Constant

shift(θ::Real)

Phase shift gate.

Definition

\[\begin{pmatrix} 1 & 0\ 0 & e^(im θ) \end{pmatrix}\]

YaoLang.GenericCircuit — Type

GenericCircuit{name}

Generic quantum circuit is the quantum counterpart of generic function.

YaoLang.Locations — Type

Locations <: AbstractLocations

Type to annotate locations in quantum circuit.

Locations(x)

Create a Locations object from a raw location statement. Valid storage types are:

Int: single position

NTuple{N, Int}: a list of locations

UnitRange{Int}: contiguous locations

Other types will be converted to the storage type via Tuple.

YaoLang.merge_locations — Method

merge_locations(locations...)

Construct a new Locations by merging two or more existing locations.

Compiler API References

YaoLang.Compiler.is_qasm_compatible — Method

is_qasm_compatible(ir)

Check if the given expression is compatible with openQASM.

YaoLang.Compiler.is_quantum — Method

is_quantum(ex)

Check if the given expression is a quantum statement.

YaoLang.Compiler.@code_yao — Macro

@code_yao <generic circuit call>

Evaluates the arguments to the function call, determines their types, and calls code_yao on the resulting expression.

YaoLang.Compiler.@ctrl — Macro

@ctrl k <gate location>

Keyword for controlled gates in quantum circuit. It must be used inside @device. See also @device.

YaoLang.Compiler.@device — Macro

@device [options] <generic circuit definition>

Entry for defining a generic quantum program. A generic quantum program is a function takes a set of classical arguments as input and return a quantum program that can be furthur compiled into pulses or other quantum instructions. The quantum program can return classical values from device if return statement is declared explicitly, or it always return nothing, and mutates the quantum register.

Supported Semantics

@ctrl: Keyword for controlled gates in quantum circuit.

@measure: Keyword for measurement in quantum circuit.

The function marked by @device can be multiple dispatched like other Julia function. The only difference is that it always returns a quantum circuit object that should be runable on quantum device by feeding it the location of qubits and the pointer to quantum register.

Options

mode, compilation mode, default is :hybrid, see Compilation Modes for details.

target, compilation target, default is :julia, see Compilation Targets for details.

Compilation Modes

:hybrid, default mode, allows arbitrary Julia code to work with quantum statements.

:pure, all classical statements are disabled.

:qasm, allows limited set of classical statements which is compatible to openQASM

Compilation Targets

:julia, default target, compiles the program to Julia program.

:qasm, compiles the program to openQASM.

Example

We can define a Quantum Fourier Transformation in the following recursive way

@device function qft(n::Int)
    1 => H
    for k in 2:n
        @ctrl k 1=>shift(2π/2^k)
    end

    if n > 1
        2:n => qft(n-1)
    end
end

This will give us a generic quantum circuit qft with 1 method.

YaoLang.Compiler.@measure — Macro

@measure <location> [operator] [configuration]

Keyword for measurement in quantum circuit. It must be used inside @device. See also @device.

Arguments

<location>: a valid Locations argument to specifiy where to measure the register

[operator]: Optional, specifiy which operator to measure

[configuration]: Optional, it can be either:

remove=true will remove the measured qubits

reset_to=<bitstring> will reset the measured qubits to given bitstring

YaoLang.Compiler.@primitive — Macro

@primitive ex

Define a primitive quantum instruction. ex can be a Symbol, if the corresponding instruction interface of YaoAPI.instruct! is implemented. Or ex can be an assignment statement for constant instructions. Or ex can be a function that returns corresponding matrix given a set of classical parameters.

Example

Since the instructions interface YaoAPI.instruct! of Pauli operators are defined, we can use

@primitive X

to declare a Pauli X primitive instruction.

Or we can also define a Hadamard primitive instruction via its matrix form

@primitive H = [1 1;1 -1]/sqrt(2)

For parameterized gates, such as phase shift gate, we can define it as

@primitive shift(θ::Real) = Diagonal([1.0, exp(im * θ)])

YaoLang.Compiler.generate_forward_stub — Method

generate_forward_stub(name::Symbol, op)

Generate forward stub which forward primitive circuit to instruction interfaces.

YaoLang.Compiler.mark_quantum — Method

mark_quantum(ir::IR)

swap the statement tag with :quantum.

YaoLang.Compiler.rm_annotations — Method

rm_annotations(x)

Remove type annotation of given expression.