Compilation

The compilation framework of YaoLang is designed to be highly extensible, so one can easily extend the semantics as long as there is no ambiguity.

For simple code generation task, one can use Julia native macros directly instead of writing a custom compiler pass. But for more complicated task, one would prefer to add a custom compiler pass.

Representation

YaoLang is a domain specific language, it embeds its own representation inside the Julia AST and SSA IR.

The Julia AST captured by a macro @device will be first transformed into custom function calls tagged with GlobalRef(YaoLang.Compiler, :node_name). The syntax check will then check if the syntax is correct, then the AST will be compiled into Julia's SSA IR as a CodeInfo object.

Then we transform the CodeInfo object to a quantum SSA IR by inferring the quantum statements, where we call this representation as YaoIR.

YaoIR

The Yao IR annotates our domain specific semantics using a :quantum head, followed by a custom head, such as :gate. For example, 1 => X will be parsed as Expr(:quantum, :gate, :X, 1).

Compiler API References

YaoLang.Compiler.YaoIR — Type

YaoIR

The Yao Intermediate Representation. See compilation section for more details.

YaoIR([m::Module=YaoLang.Compiler], ast::Expr)

Creates a YaoIR from Julia AST.

YaoLang.Compiler.is_qasm_compatible — Method

is_qasm_compatible(ir)

Check if the given expression is compatible with openQASM.

YaoLang.Compiler.is_quantum — Method

is_quantum(ex)

Check if the given expression is a quantum statement.

ZXCalculus.gate_count — Method

gate_count(circuit)::Dict

Count the number of each primitive instructions in given pure quantum circuit.

YaoLang.Compiler.@code_yao — Macro

@code_yao <generic circuit call>

Evaluates the arguments to the function call, determines their types, and calls code_yao on the resulting expression.

YaoLang.Compiler.@ctrl — Macro

@ctrl k <gate location>

Keyword for controlled gates in quantum circuit. It must be used inside @device. See also @device.

YaoLang.Compiler.@device — Macro

@device [options] <generic circuit definition>

Entry for defining a generic quantum program. A generic quantum program is a function takes a set of classical arguments as input and return a quantum program that can be furthur compiled into pulses or other quantum instructions. The quantum program can return classical values from device if return statement is declared explicitly, or it always return nothing, and mutates the quantum register.

Supported Semantics

@ctrl: Keyword for controlled gates in quantum circuit.

@measure: Keyword for measurement in quantum circuit.

The function marked by @device can be multiple dispatched like other Julia function. The only difference is that it always returns a quantum circuit object that should be runable on quantum device by feeding it the location of qubits and the pointer to quantum register.

Options

target, compilation target, default is :julia, see Compilation Targets for details.

Compilation Targets

:julia, default target, compiles the program to Julia program.

:qasm, compiles the program to openQASM.

Example

We can define a Quantum Fourier Transformation in the following recursive way

@device function qft(n::Int)
    1 => H
    for k in 2:n
        @ctrl k 1=>shift(2π/2^k)
    end

    if n > 1
        2:n => qft(n-1)
    end
end

This will give us a generic quantum circuit qft with 1 method.

YaoLang.Compiler.@measure — Macro

@measure <location> [operator] [configuration]

Keyword for measurement in quantum circuit. It must be used inside @device. See also @device.

Arguments

<location>: a valid Locations argument to specifiy where to measure the register

[operator]: Optional, specifiy which operator to measure

[configuration]: Optional, it can be either:

remove=true will remove the measured qubits

reset_to=<bitstring> will reset the measured qubits to given bitstring

YaoLang.Compiler.@primitive — Macro

@primitive ex

Define a primitive quantum instruction. ex can be a Symbol, if the corresponding instruction interface of YaoAPI.instruct! is implemented. Or ex can be an assignment statement for constant instructions. Or ex can be a function that returns corresponding matrix given a set of classical parameters.

Example

Since the instructions interface YaoAPI.instruct! of Pauli operators are defined, we can use

@primitive X

to declare a Pauli X primitive instruction.

Or we can also define a Hadamard primitive instruction via its matrix form

@primitive H = [1 1;1 -1]/sqrt(2)

For parameterized gates, such as phase shift gate, we can define it as

@primitive shift(θ::Real) = Diagonal([1.0, exp(im * θ)])

YaoLang.Compiler.@quantum — Macro

@quantum nqubits::Int <generic circuit call>

Partially evaluate all classical parts of the generic circuit.

Note

The classical part of the program could result in different circuits if it is not deterministic.

YaoLang.Compiler.generate_forward_stub — Method

generate_forward_stub(name::Symbol, op)

Generate forward stub which forward primitive circuit to instruction interfaces.

YaoLang.Compiler.mark_quantum — Method

mark_quantum(ir::IR)

swap the statement tag with :quantum.

YaoLang.Compiler.rm_annotations — Method

rm_annotations(x)

Remove type annotation of given expression.