Introduction

YaoIR is an Intermediate Representation built based on Julia builtin expression with extended semantic on quantum control, measure and position. Its (extended) syntax is very simple:

Semantics

The semantic of YaoIR tries to make use of Julia semantic as much as possible so you don't feel this is not Julian. But since the quantum circuit has some special semantic that Julia expression cannot express directly, the semantic of Julia expression is extended in YaoIR.

The point of this new IR is it make use of Julia native control flow directly instead of unroll the loop and conditions into a Julia type, such as Chain, Kron, ConditionBlock in QBIR, which improves the performance and provide possibility of further compiler optimization by analysis done on quantum circuit and classical control flows.

Gate Position

gate positions are specific with => at each line, the => operator inside function calls will not be parsed, e.g

1 => H # apply Hadamard gate on the 1st qubit
foo(1=>H) # it means normal Julia pair
1=>foo(x, y, z) # it will parse foo(x, y, z) as a quantum gate/circuit, but will error later if type inference finds they are not.

all the gate or circuit's position should be specified by its complete locations, e.g

1:n => qft(n) # right
1 => qft(n) # wrong

but single qubit gates can use multi-location argument to represent repeated locations, e.g

1:n => H # apply H on 1:n locations

Control

@ctrl is parsed as a keyword (means you cannot overload it) in each program, like QBIR, its first argument is the control location with signs as control configurations and the second argument is a normal gate position argument introduce above.

Measure

@measure is another reserved special function parsed that has specific semantic in the IR (measure the locations passed to it).

Usage

using it is pretty simple, just use @device macro to annotate a "device" function, like CUDA programming, this device function should not return anything but nothing.

The compiler will compile this function definition to a generic circuit Circuit with the same name. A generic circuit is a generic quantum program that can be overload with different Julia types, e.g

@device function qft(n::Int)
    1 => H
    for k in 2:n
        @ctrl k 1=>shift(2π/2^k)
    end

    if n > 1
        2:n => qft(n-1)
    end
end

There is no need to worry about global position: everything can be defined locally and we will infer the correct global location later either in compile time or runtime.

note: all the quantum gates should be annotate with its corresponding locations, or the compiler will not treat it as a quantum gate but instead of the original Julia expression.

Why?

There are a few reasons that we need a fully compiled DSL now.

1. Extensibility

Things in YaoBlocks like

function apply!(r::AbstractRegister, pb::PutBlock{N}) where {N}
    _check_size(r, pb)
    instruct!(r, mat_matchreg(r, pb.content), pb.locs)
    return r
end

# specialization
for G in [:X, :Y, :Z, :T, :S, :Sdag, :Tdag]
    GT = Expr(:(.), :ConstGate, QuoteNode(Symbol(G, :Gate)))
    @eval function apply!(r::AbstractRegister, pb::PutBlock{N,C,<:$GT}) where {N,C}
        _check_size(r, pb)
        instruct!(r, Val($(QuoteNode(G))), pb.locs)
        return r
    end
end

cannot be easily extended without define new dispatch on specialized instruction. Similarly, as long as there is a new instruction in low level, one need to redefine the dispatch in YaoBlocks however this is not necessary!

2. Work with classical computers

Programs defined in such way are just "normal" Julia programs, but quantum devices can be used as accelerator in a similar way comparing to GPU as an optimization.

3. More elegant and better performance

In YaoBlocks, a large quantum circuit can easily lost its structure if it is controlled, unless the programmer specialize the control block manually. Now we can map local locations into its callee location using the brand new API, thus anything in theory is composable can be executed in such way.

API References

YaoIR.H — Constant

H

The Hadamard gate.

Definition

\[\frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1\ 1 & -1 \end{pmatrix}\]

YaoIR.Rx — Constant

Rx(theta::Real)

Return a rotation gate on X axis.

YaoIR.Ry — Constant

Ry(theta::Real)

Return a rotation gate on Y axis.

YaoIR.Rz — Constant

Rz(theta::Real)

Return a rotation gate on Z axis.

YaoIR.T — Constant

T

The T gate.

YaoIR.X — Constant

X

The Pauli X gate.

YaoIR.Y — Constant

Y

The Pauli Y gate.

YaoIR.Z — Constant

Z

The Pauli Z gate.

YaoIR.phase — Constant

phase(theta)

Global phase gate.

Definition

\[exp(iθ) \mathbf{I}\]

YaoIR.rot — Constant

rot(axis, θ::T, m::Int=size(axis, 1)) where {T <: Real}

General rotation gate, axis is the rotation axis, θ is the rotation angle. m is the size of rotation space, default is the size of rotation axis.

YaoIR.shift — Constant

shift(θ::Real)

Phase shift gate.

Definition

\[\begin{pmatrix} 1 & 0\ 0 & e^(im θ) \end{pmatrix}\]

YaoIR.Locations — Type

Locations <: AbstractLocations

Type to annotate locations in quantum circuit.

Locations(x)

Create a Locations object from a raw location statement. Valid storage types are:

Int: single position

NTuple{N, Int}: a list of locations

UnitRange{Int}: contiguous locations

Other types will be converted to the storage type via Tuple.

YaoIR.merge_locations — Method

merge_locations(locations...)

Construct a new Locations by merging two or more existing locations.

YaoIR.parse_ctrl — Method

parse_ctrl(x)

Transform controlled location argument from Julia AST to Yao IR. The definition of controlled gate location is @ctrl <ctrl locations> <gate location>. The control configuration is specified using signs.

Example

quote
    @ctrl (-1, 2, 3) 4=>X
    @ctrl 1:3 4=>X
end

YaoIR.parse_locations — Method

parse_locations(x)

Transform location argument from Julia AST to Yao IR. The definition of gate locations is the first layer => in a block.

Example

quote
    1 => H # gate location
    [1=>H, 3=>X] # construct a list of pairs
    y = 1 => H # create a pair and assign it to variable y
end

YaoIR.parse_measure — Method

parse_measure(x)

Transform measurement statement from Julia AST to Yao IR. The definition of measurement is @measure <location> <operator expression>. The operator expression should return an AbstractMatrix or AbstractBlock. It will return a value which is the measurement result.

Example

quote
    c1 = @measure 1:3
    c2 = @measure 1:2 kron(X, X)
    c3 = @measure 1:2 kron(X, X) reset_to=1
end

YaoIR.@ctrl — Macro

@ctrl k <gate location>

Keyword for controlled gates in quantum circuit. It must be used inside @device. See also @device.

YaoIR.@device — Macro

@device [strict=false] <generic circuit definition>

Entry for defining a generic quantum program. A generic quantum program is a function takes a set of classical arguments as input and return a quantum circuit that can be furthur compiled into pulses or other quantum instructions.

Supported Semantics

@ctrl: Keyword for controlled gates in quantum circuit.

@measure: Keyword for measurement in quantum circuit.

The function marked by @device can be multiple dispatched like other Julia function. The only difference is that it always returns a quantum circuit object that should be runable on quantum device by feeding it the location of qubits and the pointer to quantum register.

Example

We can define a Quantum Fourier Transformation in the following recursive way

@device function qft(n::Int)
    1 => H
    for k in 2:n
        @ctrl k 1=>shift(2π/2^k)
    end

    if n > 1
        2:n => qft(n-1)
    end
end

This will give us a generic quantum circuit qft with 1 method.

YaoIR.@measure — Macro

@measure <location> [operator] [configuration]

Keyword for measurement in quantum circuit. It must be used inside @device. See also @device.

Arguments

<location>: a valid Locations argument to specifiy where to measure the register

[operator]: Optional, specifiy which operator to measure

[configuration]: Optional, it can be either:

remove=true will remove the measured qubits

reset_to=<bitstring> will reset the measured qubits to given bitstring

YaoIR.@primitive — Macro

@primitive ex

Define a primitive quantum instruction. ex can be a Symbol, if the corresponding instruction interface of YaoBase.instruct! is implemented. Or ex can be an assignment statement for constant instructions. Or ex can be a function that returns corresponding matrix given a set of classical parameters.

Example

Since the instructions interface YaoBase.instruct! of Pauli operators are defined, we can use

@primitive X

to declare a Pauli X primitive instruction.

Or we can also define a Hadamard primitive instruction via its matrix form

@primitive H = [1 1;1 -1]/sqrt(2)

For parameterized gates, such as phase shift gate, we can define it as

@primitive shift(θ::Real) = Diagonal([1.0, exp(im * θ)])

YaoIR.decode_sign — Method

decode_sign(ctrls...)

Decode signs into control sequence on control or inversed control.

YaoIR.generate_forward_stub — Method

generate_forward_stub(name::Symbol, op)

Generate forward stub which forward primitive circuit to instruction interfaces.

YaoIR.is_pure_quantum — Method

isquantum(ex)

Check if the given expression is a pure quantum circuit.

YaoIR.rm_annotations — Method

rm_annotations(x)

Remove type annotation of given expression.